+++ CHANGELOG.md
@@ -4,6 +4,8 @@
     - Grid Search: Added option to retrain best model on full dataset
     - Filesystem Persister now keeps backups of latest models
     - Added loading backup models to Persister API
+    - Added PSR-3 compatible screen logger
+    - Grid Search is now Verbose
 
 - 0.0.6-alpha
     - Added Gradient Boost regressor